Moved around files to create more organized nav bar dir
